+++ src/backend/nonfunctionalRequirements/controller.php
@@ -3,6 +3,8 @@
 require_once 'Db.php';
 require_once 'model.php';
 
+session_start();
+
 class NonfunctionalRequirementsController
 {
   private $db;
@@ -38,6 +40,40 @@ class NonfunctionalRequirementsController
     }
   }
 
+  public function fetchAllNonfunctionalRequirementsForProject($projectId): array
+  {
+    try {
+      $connection = $this->db->getConnection();
+
+      $isAuthorized = $this->isProjectOwnedByUser($projectId);
+      if (!$isAuthorized) {
+        echo "Not authorized";
+        return [];
+      }
+
+      $select = $connection->prepare(
+        'SELECT  `r`.`id`, `r`.`name`, `r`.`description`, `r`.`priority`, `r`.`project_id`, `r`.`created_at`, `nfr`.`unit`, `nfr`.`value`
+        from `requirements` `r` 
+        join `nonfunctional_requirements` `nfr` on `nfr`.`requirement_id` = `r`.`id`
+        where `r`.`project_id` = ?
+        order by `r`.`created_at` DESC'
+      );
+      $select->execute([$projectId]);
+      $rows = $select->fetchAll();
+
+      $requirements = [];
+      foreach ($rows as $row) {
+        $requirements[] = NonfunctionalRequirement::fromAssoc($row);
+      }
+
+      return $requirements;
+
+    } catch (Exception $e) {
+      echo "" . $e->getMessage();
+      return [];
+    }
+  }
+
   public function fetchNonfunctionalRequirementById($id): ?NonfunctionalRequirement
   {
     try {
@@ -51,7 +87,7 @@ class NonfunctionalRequirementsController
       );
       $select->execute([$id]);
       $requirement = $select->fetch();
-      
+
       if ($requirement === false) {
         return null;
       }
@@ -64,7 +100,12 @@ class NonfunctionalRequirementsController
 
   public function addNonfunctionalRequirement($data): bool
   {
+    $isAuthorized = $this->isProjectOwnedByUser($data['projectId']);
+    if (!$isAuthorized) {
+      return false;
+    }
     $connection = $this->db->getConnection();
+
     try {
       $connection->beginTransaction();
 
@@ -74,7 +115,7 @@ class NonfunctionalRequirementsController
       $result = $insert->execute([
         'name' => $data['name'],
         'description' => $data['description'],
-        'priority'=> $data['priority'],
+        'priority' => $data['priority'],
         'projectId' => $data['projectId']
       ]);
 
@@ -89,7 +130,7 @@ class NonfunctionalRequirementsController
         'requirementId' => $id,
       ]);
 
-      if (!$result || !$secondInsertResult ) {
+      if (!$result || !$secondInsertResult) {
         throw new Exception('Error executing INSERT statement');
       }
       $connection->commit();
@@ -105,7 +146,14 @@ class NonfunctionalRequirementsController
 
   public function removeNonfunctionalRequirementById($id): bool
   {
+    $requirement = $this->fetchNonfunctionalRequirementById($id);
+    $isAuthorized = $this->isProjectOwnedByUser($requirement->projectId);
+    if (!$isAuthorized) {
+      return false;
+    }
+
     try {
+
       $connection = $this->db->getConnection();
 
       $delete = $connection->prepare('DELETE from `requirements` `r` where `r`.`id` = ?');
@@ -114,6 +162,34 @@ class NonfunctionalRequirementsController
       return true;
     } catch (Exception $e) {
       echo "Error: " . $e->getMessage();
+      return false;
+    }
+  }
+
+  private function isProjectOwnedByUser($projectId): bool
+  {
+    try {
+      $connection = $this->db->getConnection();
+      $userId = $_SESSION['userId'];
+
+      $select = $connection->prepare(
+        'SELECT *
+        from `projects` `p` 
+        where `p`.`user_id` = :userId and `p`.`id` = :projectId'
+      );
+      $select->execute([
+        'userId' => $userId,
+        'projectId' => $projectId
+      ]);
+      $project = $select->fetch();
+
+      if ($project === false) {
+        return false;
+      }
+
+      return true;
+    } catch (Exception $e) {
+      echo "" . $e->getMessage();
       return false;
     }
   }
